pyproject.toml
[project]
name = "sd"
version = "0.1.0"
description = ""
authors = [
    {name = "Armando",email = "armandoluzborges@ufpi.edu.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin (>=6.7.0,<7.0.0)",
    "pyrebase4 (>=4.8.0,<5.0.0)",
    "pyqt6 (>=6.8.1,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "setuptools (>=76.0.0,<77.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
